refactor(cockpit): extract diff coverage gate module
